Return numpy scalar types (#1052)
Allow Numpy scalar types to be returned. Fixes #735. Add -1 to indicate the default Python precision. This allows us to differentiate between native Python types and Numpy types with the same precision (e.g. `float` and `float64`). This differentiation ensures that the final type matches the type returned by Python exactly.

**Commit Summary**
- Add `-1` to indicate the native Python precision
- Add wrapper functions to convert to Numpy scalar types
- [BUG FIX] Use `print_kind` for precision
- [BUG FIX] Map `numpy.absolute` to `numpy.abs`

+++ tests/internal/scripts/mpi/reduce.py
@@ -29,9 +29,9 @@ if __name__ == '__main__':
     root = np.int32(0)
 
     if rank == 0:
-        value = 1000
+        value = np.int32(1000)
     else:
-        value = int(rank)
+        value = rank
 
     sum_value = 0
     count = np.int32(1)
